Adding rotation augmented patch generation to preprocessing.
This adds generation of evenly-spaced rotations of mitosis patches to
the preprocessing step.  Usually, random rotations are applied in
real-time during training.  However, rotations not equal to 90 or 360
degrees will lead to the corners of the rotated patches being empty.
Although this can be filled in with mirroring, it is not ideal.  Since
we are actually extracting patches from larger images, we can instead
extract a slightly larger patch that when rotated will bound the desired
patch size in the center, without leading to empty corners.  We
implement that here in the preprocessing stage.

This also performs some minor renaming, and updates the scripts to copy
themselves to the relevant folders during preprocessing or training so
that datasets or training can be full reproduced if desired.