Adding model-bootstrapped false-positive oversampling to the data gen
This adds false-positive oversampling to the preprocessing step in which
an existing model is used to detect challenging false-positives cases
that are then oversampled while creating a new dataset.  There are a
large number of positive non-mitosis patches that could be generated,
in which a large number of them will be trivial to classify, and a small
number of them will be challenging to classify correctly.  A dataset
with uniform sampling from the normal patches will yield relatively few
challenging normal patches, and thus models trained on this dataset will
have an extremely low positive predictive value (i.e., precision).  In
order to improve performance, that model can be used to generate a new
dataset in which challenging normal patches are oversampled, hopefully
leading to higher ppv values.