Adding noise marginalization during evaluation
This adds optional noise marginalization when evaluating the validation
set.  When set, each image in the validation set will be expanded to a
batch of augmented versions of that image, and predicted probabilities
for each batch will be averaged to yield a single noise-marginalized
prediction for each image.  This is the evaluation complement of data
augmentation during training time.  The goal is to train with augmented
noise, and then marginalize over the noise at test time in order to
yield more accurate predictions.