Fixing a bug introduced during refactoring.
This fixes a bug introduced recently in which marginalization over
predictions was always applied during validation, but were then
broadcasted against the full minibatch of labels.  This caused the
metrics to be slightly incorrect, although it otherwise appeared to
the user as though all batches were being processed like normal.  In
addition to the bug fix, this introduces an assertion to prevent this
type of bug from going unnoticed.