Minor logging improvements.
The previous bug fix also identified issues with the `tf.metrics` ops in
which a race condition exists when running the metric op and the
associated metric update op at the same time.  Although the API states
that the values should be equivalent, experience shows that to not be
the case.  This fixes the issue by calling the update ops during
training or validation, and then separately calling the metric ops
(which only rely on local variables maintained with the update ops).
Additionally, this updates the logging at the end of each epoch to
output all of the metric values, and in a cleaner format.  Finally, this
moves the global epoch increment to after model checkpointing so that
the epoch number in the model checkpoint name matches that logged to
TensorBoard.