Fix random seed issue during training
It turns out that a reinitializable iterator for a Dataset will cause
any ops with random seeds, such as image augmentation ops, to be reset,
and thus each epoch will be evaluated exactly the same.  The desired
behavior would be to seed these ops once at the very beginning, so that
an entire training run can be deterministic, but not with the exact same
random augmentation during each epoch.  For now, this commit removes the
random seed for the data augmentation ops.