WIP to add multi-GPU support
This aims to add multi-GPU training support to the mitosis training
code.  In the past for the direct WSI -> tumor proliferation score
approach, we hand-rolled a multi-loss, multi-GPU algorithm that is
mathematically equivalent to a single-loss, single-GPU scenario, but it
added complexity to the code.  Before implementing that algorithm for
the mitosis approach, this commit aimed to try the `multi_gpu_model`
function that was added to Keras.  Unfortunately, we have experienced
severe performance issues with that function so far, and due to that, we
also have attempted to make adjustments to that function.  So far, we
are still experiencing severe performance issues, so this addition is
still a work in progress, and may be replaced with the multi-loss,
multi-GPU approach, or with an entirely different approach.