Improve logging of command line arguments for experiments & datasets
Previously, we wrote out the `Namespace` arguments object to a file in
order to keep track of the arguments used for a training experiment or
for a dataset.  This became difficult to use later, so this change
instead writes out a json representation of the arguments.
Additionally, we write out a file with the full command line invocation,
aside from any environment variables such as `CUDA_VISIBLE_DEVICES`.
Together, this will improve the ability to analyze hyperparameters
(i.e., evaluate a hyperparam search) and re-run experiments or
hyperparameter searches.