Force computation of the RDD before converting it to a DataFrame
This forces computation of an RDD in the mitosis prediction code path
before converting it to a DataFrame in other to maximize parallelism.
Otherwise, the parallelism will be initially reduced to a single task to
check the schema of the resulting DataFrame.  The prediction jobs are
too computationally intensive to waste time on such things, so by
executing and caching the RDD, we can maximize the amount of parallel
processing, and thus spend more time on research.