Only filter partial ROIs during prediction if using ROIs
In the prediction code, we previously would filter out partial ROIs,
such as a part of an ROI at the right side or bottom of a slide, so that
we would only consider full-sized ROIs.  When we used the prediction
code for mitosis prediction and thus skipped ROIs, we would still apply
this filter.  This became problematic for the ICPR datasets that have
images for which the height and width are not equal. This fixes that
issue by only applying the filter if we are using ROIs.

Additionally, this updates some docstrings.