Move rotation & translation augmentation into the training code
This implements the random rotation and random translation augmentation
from the preprocessing code in the training augmentation code so that
these methods can be performed in real-time.  The goal is to expand the
amount of augmentation by adding all of it into training code.

In order to make this work properly, this code ideally assumes that the
preprocessed patches will be large enough to be able to be rotated to
the worst case of 45 degrees and still have an inner area large enough
to extract a centered patch without empty corners that will be large
enough for random translation via cropping to a final patch size.  I.e.,
we want to perform random rotation and random translation without
introducing whitespace into the corners of the image, so we need to
create larger preprocessed patches before training.

The random translation is implemented via random cropping in such a way
that the Euclidean distance to the true center of the mitotic figure is
guaranteed to be <= 26 pixels, or 6.5 µm, which gives a slight amount
of buffer to the upper bound of 30 pixels (7.5 µm) that is allowed for a
positive prediction.